requirements.txt
pandas==1.1.3
torchvision==0.9.0
selenium==3.141.0
detectron2==0.4
opencv_python_headless==4.5.1.48
numpy==1.19.2
beautifulsoup4==4.10.0
Pillow==8.3.2
